+++ src/lib/form-field/form-field.component.ts
@@ -1,4 +1,4 @@
-import {Component, ComponentFactoryResolver, OnChanges, SimpleChanges, ViewContainerRef} from '@angular/core';
+import {Component, ComponentFactoryResolver, OnChanges, SimpleChange, SimpleChanges, ViewContainerRef} from '@angular/core';
 import {AbstractFormFieldComponent} from '../abstract-form-field/abstract-form-field.component';
 import {AbstractGroupType} from '../form-type/abstract-group-type';
 import {FormGroup} from '@angular/forms';
@@ -31,12 +31,26 @@ export class FormFieldComponent extends AbstractFormFieldComponent<any> implemen
     const component = this.viewRef.createComponent(factory);
 
     const isGroup = this.fieldType instanceof AbstractGroupType;
-    component.instance.formGroup = isGroup ? this.element as FormGroup : this.formGroup;
-    component.instance.element = this.element;
-    component.instance.fieldType = this.fieldType;
-    component.instance.slots = this.slots;
-    component.instance.path = this.path;
-    component.instance.index = this.index;
+    const changes = {
+      formGroup: new SimpleChange(null, isGroup ? this.element as FormGroup : this.formGroup, true),
+      element: new SimpleChange(null, this.element, true),
+      fieldType: new SimpleChange(null, this.fieldType, true),
+      slots: new SimpleChange(null, this.slots, true),
+      path: new SimpleChange(null, this.path, true),
+      index: new SimpleChange(null, this.index, true),
+    };
+
+    component.instance.formGroup = changes.formGroup.currentValue;
+    component.instance.element = changes.element.currentValue;
+    component.instance.fieldType = changes.fieldType.currentValue;
+    component.instance.slots = changes.slots.currentValue;
+    component.instance.path = changes.path.currentValue;
+    component.instance.index = changes.index.currentValue;
+
+    if ('ngOnChanges' in component.instance) {
+      (component.instance as OnChanges).ngOnChanges(changes);
+    }
+
     component.changeDetectorRef.detectChanges();
   }
 }