+++ src/lib/form-type/abstract-group-type.ts
@@ -1,5 +1,4 @@
-import {AbstractType} from './abstract-type';
-import {FormModel} from '../form-builder/form-builder.component';
+import {AbstractType, FormModel} from './abstract-type';
 import {FormGroup} from '../abstraction';
 
 export interface IGroupTypeOptions {
@@ -7,7 +6,5 @@ export interface IGroupTypeOptions {
 }
 
 export abstract class AbstractGroupType<T extends IGroupTypeOptions> extends AbstractType<T> {
-
   public control: FormGroup;
-
 }