+++ src/lib/form-type/abstract-layout-type.ts
@@ -1,0 +1,4 @@
+import {AbstractGroupType, IGroupTypeOptions} from './abstract-group-type';
+
+export abstract class AbstractLayoutType<T extends IGroupTypeOptions> extends AbstractGroupType<T> {
+}